Adding and moving tests
